requirements.txt
# Core LLM & Agent Framework
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0

# Vector DB
chromadb>=0.5.0

# Embeddings (local, no API key needed)
sentence-transformers>=3.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web UI
streamlit>=1.35.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0